Revert "5.4  最后修改合央视开始修改"
This reverts commit 29658c159d5ff39964d6083e17e69685b400e2e6.

+++ yixiupige/MODEL/JCInfoModel.cs
@@ -8,7 +8,7 @@ namespace MODEL
 {
     public class JCInfoModel
     {
-        public double jcNo { get; set; }
+        public int jcNo { get; set; }
         public int jcID { get; set; }
         public string jcCardNumber { get; set; }
         public string jcName { get; set; }
